+++ beatmap/v3/slider.ts
@@ -1,5 +1,5 @@
 import { BaseSlider } from './baseSlider.ts';
-import { LINE_COUNT, NoteCutAngle } from '../shared/constants.ts';
+import { LINE_COUNT, NoteDirectionAngle } from '../shared/constants.ts';
 import { ISlider } from '../../types/beatmap/v3/slider.ts';
 import { ObjectReturnFn } from '../../types/utils.ts';
 import { deepCopy } from '../../utils/misc.ts';
@@ -251,19 +251,19 @@ export class Slider extends BaseSlider<ISlider> {
     getTailAngle(type?: 'vanilla' | 'me' | 'ne') {
         switch (type) {
             case 'vanilla':
-                return NoteCutAngle[this.tailDirection as keyof typeof NoteCutAngle] || 0;
+                return NoteDirectionAngle[this.tailDirection as keyof typeof NoteDirectionAngle] || 0;
             case 'me':
                 if (this.tailDirection >= 1000) {
                     return Math.abs(((this.tailDirection % 1000) % 360) - 360);
                 }
             /* falls through */
             case 'ne':
-                return NoteCutAngle[this.tailDirection as keyof typeof NoteCutAngle] || 0;
+                return NoteDirectionAngle[this.tailDirection as keyof typeof NoteDirectionAngle] || 0;
             default:
                 if (this.tailDirection >= 1000) {
                     return Math.abs(((this.tailDirection % 1000) % 360) - 360);
                 }
-                return NoteCutAngle[this.tailDirection as keyof typeof NoteCutAngle] || 0;
+                return NoteDirectionAngle[this.tailDirection as keyof typeof NoteDirectionAngle] || 0;
         }
     }